// src/types/objective-types.ts
export type HelpBadge = {
  path: string;
  content: string;
  color: string;
};

export enum ObjectiveType {
  POLICY_CREATION_OBJECTIVE = 'POLICY_CREATION_OBJECTIVE',
  RESOURCE_POLICY_CREATION_OBJECTIVE = 'RESOURCE_POLICY_CREATION_OBJECTIVE',
  SCP_CREATION_OBJECTIVE = 'SCP_CREATION_OBJECTIVE',
  POLICY_EDIT_OBJECTIVE = 'POLICY_EDIT_OBJECTIVE',
  TRUST_POLICY_EDIT_OBJECTIVE = 'TRUST_POLICY_EDIT_OBJECTIVE',
  ROLE_CREATION_OBJECTIVE = 'ROLE_CREATION_OBJECTIVE',
  IAM_USER_GROUP_CREATION_OBJECTIVE = 'IAM_USER_GROUP_CREATION_OBJECTIVE',
  EDGE_CONNECTION_OBJECTIVE = 'EDGE_CONNECTION_OBJECTIVE',
  LEVEL_OBJECTIVE = 'LEVEL_OBJECTIVE',
  PERMISSION_BOUNDARY_CREATION_OBJECTIVE = 'PERMISSION_BOUNDARY_CREATION_OBJECTIVE',
}

export type BaseFinishEventMap = Record<ObjectiveType, string>;
// ...
export interface BaseCreationObjective<TFinishEventMap extends BaseFinishEventMap> {
  /**
   * Unique identifier for the objective.
   * This ID is used as the node ID upon creation and as the objective's identifier in the state machine context.
   */
  readonly id: string;
  readonly entity: unknown;
  readonly type: ObjectiveType;
  readonly initial_code: object;
  readonly on_finish_event:
    | TFinishEventMap[ObjectiveType.POLICY_CREATION_OBJECTIVE]
    | TFinishEventMap[ObjectiveType.SCP_CREATION_OBJECTIVE]
    | TFinishEventMap[ObjectiveType.ROLE_CREATION_OBJECTIVE]
    | TFinishEventMap[ObjectiveType.RESOURCE_POLICY_CREATION_OBJECTIVE]
    | TFinishEventMap[ObjectiveType.PERMISSION_BOUNDARY_CREATION_OBJECTIVE];
  readonly help_badges?: HelpBadge[];
  readonly limit_new_lines?: boolean;
  readonly initial_position?: string;
  readonly account_id?: string;
  readonly created_node_initial_position?: string;
  readonly callout_message?: string;
  readonly hint_messages?: { title: string; content: string }[];
  readonly created_node_parent_id?: string;
  readonly layout_group_id?: string;
  readonly initial_edges?: {
    from: string;
    to: string;
  }[];
  finished: boolean;
  /**
   * Extra data overridable by each objective type
   */
  extra_data: Record<string, unknown>;
  /**
   * Optional tooltip to show on the created node via a pulsing dot indicator.
   */
  node_tooltip?: string;
}
// ...
export interface IAMResourcePolicyCreationObjective<
  TFinishEventMap extends BaseFinishEventMap,
  TApplicableNodesFnName extends string = string,
> extends BaseCreationObjective<TFinishEventMap> {
  readonly entity: IAMNodeEntity.ResourcePolicy;
  readonly type: ObjectiveType.RESOURCE_POLICY_CREATION_OBJECTIVE;
  readonly on_finish_event: TFinishEventMap[ObjectiveType.RESOURCE_POLICY_CREATION_OBJECTIVE];
  readonly extra_data: {
    readonly resource_node_id: string;
    readonly granted_accesses?: PolicyGrantedAccess<TApplicableNodesFnName>[];
  };
}